test: move subscription tests into accounts test suite and remove legacy tests
